requirements.txt
transformers>=4.37.0
torch>=2.2.0
sentence-transformers>=2.3.1
chromadb>=0.4.22
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
numpy>=1.26.0
requests>=2.31.0
streamlit>=1.31.0
